Fix():Remouving warnings
