+++ crates/cloud-wallet-openid4vc/src/oid4vp/presentation/mod.rs
@@ -1,11 +1,15 @@
 mod builder;
+mod error;
 mod holder_binding;
 mod vp_token;
 
-pub use builder::{PresentationBuilder, PresentationBuilderError, SelectedCredential};
+pub use builder::{PresentationBuilder, SelectedCredential};
+pub use error::{
+    PresentationBuilderError, HolderBindingProofError, VpTokenBuilderError,
+};
 pub use holder_binding::{
-    HolderBinding, HolderBindingProof, HolderBindingProofError, HolderBindingFormat,
+    HolderBinding, HolderBindingProof, HolderBindingFormat,
     KeyBindingInput, KeyBindingClaims, SdJwtHolderBinding, MdocHolderBinding,
     compute_sd_hash, build_key_binding_jwt_claims,
 };
-pub use vp_token::{VpTokenBuilder, VpTokenBuilderError, VpTokenResponse};
+pub use vp_token::{VpTokenBuilder, VpTokenResponse};
